#221 Adds IGDB Image Fields Facade
